+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 
   Thanks to Renan Rodrigues in `PR #344 <https://github.com/adamchainz/ec2-metadata/pull/344>`__.
 
+* Add support for ``spot/instance-action``.
+
 2.7.0 (2022-01-10)
 ------------------